+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-04 22:03:36
-// Commit: fix(core): resolve race condition during core initialization
+// Updated at: 2026-08-06 09:40:43
+// Commit: feat(core): integrate custom telemetry logger in core
 
 class coreComponentState {
-  final String commitRef = '3371';
-  final String lastMessage = 'fix(core): resolve race condition during core initialization';
-  final DateTime timestamp = DateTime.parse('2026-08-04T22:03:36+05:00');
+  final String commitRef = '3377';
+  final String lastMessage = 'feat(core): integrate custom telemetry logger in core';
+  final DateTime timestamp = DateTime.parse('2026-08-06T09:40:43+05:00');
 }